+++ projects/vedana/CHANGELOG.md
@@ -1,3 +1,7 @@
+# WIP
+
+* (vedana) Turn off built-in password auth
+
 # 2025.07.21-dev.8
 
 * (vedana) Translate UI fields and pipeline logs in English